Put declaration of Multi_Index container into a new header
