+++ src/libNode/Node.h
@@ -16,11 +16,6 @@
 #ifndef __NODE_H__
 #define __NODE_H__
 
-#include <boost/multi_index/composite_key.hpp>
-#include <boost/multi_index/hashed_index.hpp>
-#include <boost/multi_index/key_extractors.hpp>
-#include <boost/multi_index/ordered_index.hpp>
-#include <boost/multi_index_container.hpp>
 #include <condition_variable>
 #include <deque>
 #include <list>
@@ -33,6 +28,7 @@
 #include "common/Broadcastable.h"
 #include "common/Constants.h"
 #include "common/Executable.h"
+#include "common/MultiIndexContainer.h"
 #include "depends/common/FixedHash.h"
 #include "libConsensus/Consensus.h"
 #include "libData/AccountData/Transaction.h"
@@ -46,40 +42,6 @@
 
 class Mediator;
 class Retriever;
-
-typedef boost::multi_index::composite_key<
-    Transaction,
-    boost::multi_index::const_mem_fun<Transaction, Address,
-                                      &Transaction::GetSenderAddr>,
-    boost::multi_index::const_mem_fun<Transaction,
-                                      const boost::multiprecision::uint256_t&,
-                                      &Transaction::GetNonce>>
-    addr_nonce_key;
-
-namespace boost
-{
-    namespace multiprecision
-    {
-        static std::size_t hash_value(const uint256_t& value)
-        {
-            std::size_t seed = 0;
-            boost::hash_combine(seed, value);
-            return seed;
-        }
-    }
-}
-
-typedef boost::multi_index::multi_index_container<
-    Transaction,
-    boost::multi_index::indexed_by<
-        boost::multi_index::ordered_non_unique<
-            boost::multi_index::const_mem_fun<
-                Transaction, const boost::multiprecision::uint256_t&,
-                &Transaction::GetGasPrice>>,
-        boost::multi_index::hashed_unique<boost::multi_index::const_mem_fun<
-            Transaction, const TxnHash&, &Transaction::GetTranID>>,
-        boost::multi_index::hashed_unique<addr_nonce_key>>>
-    gas_ra_txns;
 
 /// Implements PoW submission and sharding node functionality.
 class Node : public Executable, public Broadcastable
@@ -186,7 +148,7 @@ class Node : public Executable, public Broadcastable
 
     // Transactions information
     std::mutex m_mutexCreatedTransactions;
-    gas_ra_txns m_createdTransactions;
+    gas_txnid_comp_txns m_createdTransactions;
 
     std::unordered_map<Address,
                        std::map<boost::multiprecision::uint256_t, Transaction>>
